+++ DMSUI.Services/Interfaces/ICalibrationManager.cs
@@ -14,7 +14,7 @@ namespace DMSUI.Services.Interfaces
 		Task<CalibrationItemDTO> GetCalibrationItemByIdAsync(int calibrationId);
 		Task<ulong?> CreateCalibrationAsync(CreateCalibrationDTO createCalibrationDTO);
 
-		Task<bool> UpdateCalibrationAsync(ulong id, EditCalibrationDTO editCalibrationDTO);
+		Task<bool> UpdateCalibrationAsync(int id, EditCalibrationDTO editCalibrationDTO);
 
 		Task<bool> DeleteByIdAsync(int calibrationId);
 	}
